+++ core/common/program_class.ts
@@ -12,6 +12,7 @@ export interface ModulesFunctions {
     programName: string,
   ) => void;
   initializer?: InitializerFunction;
+  getState?: () => Record<string, unknown>;
 }
 export interface ShaderPart {
   useWith: (name: string) => void;